fix: Make sure to test for automated-popup-opening + set proper popup-policy again
AB#32738
