+++ interfaces/Portalicious/_build-deployment-configuration.mjs
@@ -100,7 +100,8 @@ if (
     '✅ Allow control of pop-ups for SSO when the Portal is in an iframe on Twilio Flex',
   );
 
-  swaConfig.globalHeaders['Cross-Origin-Opener-Policy'] = 'unsafe-none';
+  swaConfig.globalHeaders['Cross-Origin-Opener-Policy'] =
+    'same-origin-allow-popups';
 }
 
 // Feature: PowerBI Dashboard(s)
