+++ interfaces/Portalicious/_test-deployment-configuration.mjs
@@ -62,7 +62,7 @@ test('Configuration to control pop-ups for SSO when the Portal is in an iframe o
     process.env.USE_IN_TWILIO_FLEX_IFRAME === 'true' &&
     process.env.USE_SSO_AZURE_ENTRA === 'true'
   ) {
-    match(openerPolicy, /unsafe-none/);
+    match(openerPolicy, /same-origin-allow-popups/);
   } else {
     match(openerPolicy, /same-origin/);
   }
