+++ interfaces/Portalicious/src/app/services/auth/strategies/msal-auth/msal-auth.login.component.ts
@@ -20,9 +20,9 @@ import { ProgressSpinnerModule } from 'primeng/progressspinner';
 import { FormDefaultComponent } from '~/components/form/form-default.component';
 import { FormFieldWrapperComponent } from '~/components/form-field-wrapper/form-field-wrapper.component';
 import { AuthService } from '~/services/auth.service';
+import { areAutomatedPopupsBlocked } from '~/utils/are-popups-blocked';
 import { generateFieldErrors } from '~/utils/form-validation';
 import { isIframed } from '~/utils/is-iframed';
-import { isPopupBlocked } from '~/utils/is-pop-up-blocked';
 
 type LoginFormSsoGroup =
   (typeof MsalAuthLoginComponent)['prototype']['formGroup'];
@@ -46,6 +46,8 @@ type LoginFormSsoGroup =
 })
 export class MsalAuthLoginComponent {
   private authService = inject(AuthService);
+  private arePopupsAutoBlocked = areAutomatedPopupsBlocked(); // Should not be called in a user-initiated event
+
   returnUrl = input<string | undefined>(undefined);
 
   formGroup = new FormGroup({
@@ -63,9 +65,10 @@ export class MsalAuthLoginComponent {
       return $localize`Enter a valid email address`;
     },
   });
+
   loginMutation = injectMutation(() => ({
     onMutate: () => {
-      if (isIframed() && isPopupBlocked()) {
+      if (isIframed() && this.arePopupsAutoBlocked) {
         throw Error(
           $localize`Please allow pop-up windows in your browser settings to login`,
         );
